# labels: faster Compare function when using -tags stringlabels 