Instead of unpacking every individual string, we skip to the point where there is a difference.

Added a benchmark to illustrate.

[benchmark results updated 2023-06-15T1513]
Before/after this PR with `-tags stringlabels`:
```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
                                 │  before.txt  │             after.txt              │
                                 │    sec/op    │   sec/op     vs base               │
Labels_Compare/equal-4              71.13n ± 6%   13.40n ± 2%  -81.16% (p=0.002 n=6)
Labels_Compare/not_equal-4          69.54n ± 2%   33.32n ± 2%  -52.09% (p=0.002 n=6)
Labels_Compare/different_sizes-4   39.050n ± 2%   9.840n ± 4%  -74.80% (p=0.002 n=6)
Labels_Compare/lots-4              300.85n ± 9%   46.59n ± 4%  -84.51% (p=0.002 n=6)
geomean                             87.31n        21.27n       -75.64%
```

After this PR, without/with `-tags stringlabels`:
```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
                                 │  slice.txt   │             after.txt              │
                                 │    sec/op    │   sec/op     vs base               │
Labels_Compare/equal-4              23.25n ± 3%   13.40n ± 2%  -42.35% (p=0.002 n=6)
Labels_Compare/not_equal-4          25.02n ± 7%   33.32n ± 2%  +33.20% (p=0.002 n=6)
Labels_Compare/different_sizes-4   14.365n ± 6%   9.840n ± 4%  -31.50% (p=0.002 n=6)
Labels_Compare/lots-4               92.77n ± 3%   46.59n ± 4%  -49.77% (p=0.002 n=6)
geomean                             29.67n        21.27n       -28.31%
```